+++ src/commands/announce.cpp
@@ -1,0 +1,9 @@
+#include "common.h"
+
+#include "announce.h"
+
+void
+cmd_announce(CSTRING data)
+{
+	UNUSED_PARAM(data);
+}