+++ src/commands/announce.h
@@ -1,0 +1,8 @@
+#ifndef SRC_COMMANDS_ANNOUNCE_H_
+#define SRC_COMMANDS_ANNOUNCE_H_
+
+__SWIRC_BEGIN_DECLS
+void	cmd_announce(CSTRING);
+__SWIRC_END_DECLS
+
+#endif