+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased] ##
+- Added command `/announce`.
 - Added event `CHGHOST` (IRCv3.)
 - Added option `chghost` (bool). Send a notification when clients
   change their username or host. Defaults to **yes**.